TEST: consolidate shared Result test helpers and fixtures
